+++ rules/ecmascript6.js
@@ -59,7 +59,8 @@ module.exports = {
     // Disallow `parseInt()` in favor of binary, octal, and hexadecimal literals
     "prefer-numeric-literals" : "error",
     // Suggest using Reflect methods where applicable
-    "prefer-reflect" : "error",
+    // **deprecated** in ESLint v3.9.0
+    "prefer-reflect" : "off",
     // Suggest using the rest parameters instead of arguments
     "prefer-rest-params" : "error",
     // Suggest using the spread operator instead of .apply().